restaurante2/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cryptography==45.0.2
defusedxml==0.7.1
Django==5.2
django-allauth==0.61.0
django-widget-tweaks==1.5.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python3-openid==3.2.0
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
